pyproject.toml
[project]
name = "playwright-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromium>=0.0.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-order>=1.3.0",
    "pytest-playwright>=0.7.2",
    "pytest-xdist>=3.8.0",
]
